Updated Render Components
Inheritance from Render Component was removed from TileMapComponent and
ImageComponent. Both component had a layer field added.
Render Component is still present though renamed "RenderableComponent"
with a single flag to signify active and now acts as a "Render This"
tag. The next iteration of the Render System will only act on entities
with a Renderable Tag set to True.